Updated Utils Function with Header-File
